nextfoodafrica/requirements.txt
absl-py==2.0.0
apturl==0.5.2
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
astunparse==1.6.3
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.1.0
cachetools==5.3.2
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
command-not-found==0.3
cryptography==42.0.2
cssselect2==0.7.0
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
dj-database-url==2.1.0
Django==3.0.12
django-autoslug==1.9.9
django-ckeditor==6.7.1
django-crispy-forms==2.1
django-heroku==0.3.1
django-js-asset==2.2.0
django-widget-tweaks==1.5.0
duplicity==0.8.21
fasteners==0.14.1
filelock==3.15.4
flatbuffers==23.5.26
fonttools==4.53.1
fsspec==2024.6.1
future==0.18.2
gast==0.5.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
gpg==1.16.0
greenlet==3.0.1
grpcio==1.59.3
h5py==3.10.0
html5lib==1.1
httplib2==0.20.2
huggingface-hub==0.24.5
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
joblib==1.3.2
keras==2.15.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==16.0.6
lockfile==0.12.2
louis==3.20.0
lxml==5.1.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.5.1
MarkupSafe==2.1.3
ml-dtypes==0.2.0
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.26.2
oauthlib==3.2.0
olefile==0.46
opt-einsum==3.3.0
oscrypto==1.3.0
packaging==23.2
pandas==2.1.3
paramiko==2.9.3
pexpect==4.8.0
pillow==10.4.0
polars==0.19.16
protobuf==4.23.4
psycopg2==2.8.6
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydyf==0.11.0
PyGObject==3.42.1
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pypdf==4.0.1
pyphen==0.16.0
pypng==0.20220715.0
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-bidi==0.4.2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
pytz==2022.1
pyxdg==0.27
PyYAML==6.0.1
qrcode==7.4.2
razorpay==1.4.1
regex==2024.7.24
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.4
scikit-learn==1.3.2
scipy==1.11.4
SecretStorage==3.3.1
six==1.16.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
ssh-import-id==5.11
svglib==1.5.1
systemd-python==234
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
threadpoolctl==3.2.0
tinycss2==1.3.0
tokenizers==0.19.1
tqdm==4.66.5
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritools==4.0.2
urllib3==1.26.5
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.6
weasyprint==62.3
webencodings==0.5.1
Werkzeug==3.0.1
whitenoise==6.6.0
wrapt==1.14.1
wxPython==4.0.7
xdg==5
xhtml2pdf==0.2.14
xkit==0.0.0
zipp==1.0.0
zopfli==0.2.3